feat: account service hooks
